Refactor models and constants file to models folder
